+++ nflow-rest-api-common/src/test/java/io/nflow/rest/v1/converter/ListWorkflowDefinitionConverterTest.java
@@ -53,7 +53,7 @@ public class ListWorkflowDefinitionConverterTest {
     assertThat((int)resp.settings.transitionDelaysInMilliseconds.minErrorWait, is(def.getSettings().minErrorTransitionDelay));
     assertThat((int)resp.settings.transitionDelaysInMilliseconds.maxErrorWait, is(def.getSettings().maxErrorTransitionDelay));
     assertThat(resp.settings.maxRetries, is(def.getSettings().maxRetries));
-    assertThat(resp.settings.historyDeletableAfterHours, is(def.getSettings().historyDeletableAfterHours));
+    assertThat(resp.settings.historyDeletableAfter, is(def.getSettings().historyDeletableAfter));
   }
 
   private State getResponseState(DummyTestWorkflow.State workflowState, List<String> nextStateNames, String errorStateName) {